update readme and refactor classes
