+++ classes/hea_alloy.py
@@ -1,4 +1,4 @@
-from transformer import Transformer
+from classes.transformer import Transformer
 from services.logger import logger
 
 
@@ -6,13 +6,13 @@ class HEA_Alloy:
     
     def __init__ (self, row : str) -> None :
         
-        self._composition: list[ float ] = [0] * len(Transformer.elements)
-        self._elements_dict : dict[str: float] = Transformer.composition_to_molar_elements(row) 
-        self._composition_str : str = row
+        self._composition: list[ float ]        = [0] * len(Transformer.elements)
+        self._elements_dict : dict[str: float]  = Transformer.composition_to_molar_elements(row) 
+        self._composition_str : str             = row
 
         # fill in composition with molars from row data
         elements = list(self._elements_dict.keys())
-        logger.debug(f'Created HEA_Alloy with elements: {elements}')
+        logger.debug(f'Created a HEA_Alloy with elements: {elements}')
         molars   = Transformer.normalize_molar_ratios(list(self._elements_dict.values()))
         for j, el in enumerate(elements):
             el_idx = Transformer.elements.index(el)
